* next:
  minor, use standard main() signature
  fix, use bitwise & instead of logical &&
  minor, use -w2 if possible
  minor, obey icc warnings
  minor, avoid declaring FILE* for ud_mod_dump(), turn into a void*
  minor, add user's CFLAGS not autoconf's
  minor, make configuration a bit leaner
  minor, kick static intel, add clang and tcc

+++ src/tcp-unix.c
@@ -404,7 +404,7 @@ data_cb(EV_P_ ev_io *w, int UNUSED(re))
 		goto clo;
 	}
 	UD_DEBUG_TU("new data in sock %d\n", w->fd);
-	if (nrd < sizeof(buf)) {
+	if ((size_t)nrd < sizeof(buf)) {
 		/* just a service so that message parsers can use the string */
 		buf[nrd] = '\0';
 	}
